[*] refactoring
